refactor: move `processLabelMatcher` to shared PromQL utils
