+++ src/extensions/links.ts
@@ -14,6 +14,7 @@ import { parseMatcher } from 'WingmanDataTrail/MetricVizPanel/parseMatcher';
 
 import { PLUGIN_BASE_URL, ROUTES } from '../constants';
 import { logger } from '../tracking/logger/logger';
+import { processLabelMatcher, type ParsedPromQLQuery, type PromQLLabelMatcher } from '../utils/utils.promql';
 
 const PRODUCT_NAME = 'Grafana Metrics Drilldown';
 const title = `Open in ${PRODUCT_NAME}`;
@@ -25,12 +26,6 @@ export const ASSISTANT_TARGET_V0 = 'grafana-metricsdrilldown-app/grafana-assista
 export const ASSISTANT_TARGET_V1 = 'grafana-assistant-app/navigateToDrilldown/v1';
 
 export const ADHOC_URL_DELIMITER = '|';
-
-export type PromQLLabelMatcher = {
-  label: string;
-  op: string;
-  value: string;
-};
 
 export const linkConfigs: Array<PluginExtensionAddedLinkConfig<PluginExtensionPanelContext>> = [
   {
@@ -138,13 +133,6 @@ export function configureDrilldownLink<T extends PluginExtensionPanelContext>(co
   }
 }
 
-export interface ParsedPromQLQuery {
-  metric: string;
-  labels: PromQLLabelMatcher[];
-  hasErrors: boolean;
-  errors: string[];
-}
-
 export function parsePromQLQuery(expr: string): ParsedPromQLQuery {
   const tree = parser.parse(expr);
   let metric = '';
@@ -179,35 +167,6 @@ export function parsePromQLQuery(expr: string): ParsedPromQLQuery {
   });
 
   return { metric, labels, hasErrors, errors };
-}
-
-// Helper function to process label matcher nodes
-export function processLabelMatcher(node: any, expr: string): PromQLLabelMatcher | null {
-  if (node.name !== 'UnquotedLabelMatcher') {
-    return null;
-  }
-
-  const labelNode = node.node;
-  let labelName = '';
-  let op = '';
-  let value = '';
-
-  // Get children of UnquotedLabelMatcher
-  for (let child = labelNode.firstChild; child; child = child.nextSibling) {
-    if (child.type.name === 'LabelName') {
-      labelName = expr.slice(child.from, child.to);
-    } else if (child.type.name === 'MatchOp') {
-      op = expr.slice(child.from, child.to);
-    } else if (child.type.name === 'StringLiteral') {
-      value = expr.slice(child.from + 1, child.to - 1); // Remove quotes
-    }
-  }
-
-  if (labelName && op) {
-    // Allow empty string values
-    return { label: labelName, op, value };
-  }
-  return null;
 }
 
 /**
@@ -325,7 +284,10 @@ function isPromQuery(query: DataQuery): query is PromQuery {
 
 // Copied from interpolateQueryExpr in prometheus datasource, as we can't return a promise in the link extension config we can't fetch the datasource from the datasource srv, so we're forced to duplicate this method
 // eslint-disable-next-line sonarjs/function-return-type
-export function interpolateQueryExpr(value: string | string[] = [], variable: QueryVariableModel | CustomVariableModel): string | string[] {
+export function interpolateQueryExpr(
+  value: string | string[] = [],
+  variable: QueryVariableModel | CustomVariableModel
+): string | string[] {
   // if no multi or include all do not regexEscape
   if (!variable.multi && !variable.includeAll) {
     return prometheusRegularEscape(value);
