+++ src/utils/utils.promql.ts
@@ -1,5 +1,18 @@
 import { parser } from '@prometheus-io/lezer-promql';
 
+export interface PromQLLabelMatcher {
+  label: string;
+  op: string;
+  value: string;
+}
+
+export interface ParsedPromQLQuery {
+  metric: string;
+  labels: PromQLLabelMatcher[];
+  hasErrors: boolean;
+  errors: string[];
+}
+
 /**
  * Extracts all metric names from a PromQL expression
  * @param {string} promqlExpression - The PromQL expression to parse
@@ -35,4 +48,33 @@ export function extractMetricNames(promqlExpression: string): string[] {
   } while (cursor.next());
 
   return Array.from(metricNames);
+}
+
+// Helper function to process label matcher nodes
+export function processLabelMatcher(node: any, expr: string): PromQLLabelMatcher | null {
+  if (node.name !== 'UnquotedLabelMatcher') {
+    return null;
+  }
+
+  const labelNode = node.node;
+  let labelName = '';
+  let op = '';
+  let value = '';
+
+  // Get children of UnquotedLabelMatcher
+  for (let child = labelNode.firstChild; child; child = child.nextSibling) {
+    if (child.type.name === 'LabelName') {
+      labelName = expr.slice(child.from, child.to);
+    } else if (child.type.name === 'MatchOp') {
+      op = expr.slice(child.from, child.to);
+    } else if (child.type.name === 'StringLiteral') {
+      value = expr.slice(child.from + 1, child.to - 1); // Remove quotes
+    }
+  }
+
+  if (labelName && op) {
+    // Allow empty string values
+    return { label: labelName, op, value };
+  }
+  return null;
 }
